website/requirements.txt
Django==2.0.2
django-admin-interface==0.6.3
django-environ==0.4.4
